pyproject.toml
[project]
name = "wb-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.18.0",
    "alembic>=1.15.1",
    "apscheduler>=3.11.0",
    "asyncpg>=0.30.0",
    "black>=25.1.0",
    "fastapi>=0.115.11",
    "loguru>=0.7.3",
    "pydantic-settings>=2.8.1",
    "pytz>=2025.1",
    "sqlalchemy>=2.0.39",
    "uvicorn>=0.34.0",
]
